+++ lib/src/autoimpl.rs
@@ -13,7 +13,8 @@ use quote::{quote, TokenStreamExt};
 use syn::spanned::Spanned;
 use syn::token::Comma;
 use syn::{
-    parse2, Field, Fields, Ident, Index, Item, ItemStruct, Member, Path, PathArguments, Token,
+    parse2, Field, Fields, Ident, Index, Item, ItemEnum, ItemStruct, Member, Path, PathArguments,
+    Token,
 };
 
 mod impl_misc;
@@ -74,6 +75,28 @@ pub trait ImplTrait {
         false
     }
 
+    /// Generate an impl for an enum item
+    ///
+    /// The default implementation is a wrapper around [`Self::enum_items`]
+    /// and suffices for most cases. It may be overridden, e.g. to generate
+    /// multiple implementation items. It is not recommended to modify the
+    /// generics.
+    fn enum_impl(&self, item: &ItemEnum, args: &ImplArgs) -> Result<Toks> {
+        let type_ident = &item.ident;
+        let (impl_generics, ty_generics, item_wc) = item.generics.split_for_impl();
+
+        let (path, items) = self.enum_items(item, args)?;
+
+        let wc = clause_to_toks(&args.clause, item_wc, &path);
+
+        Ok(quote! {
+            #[automatically_derived]
+            impl #impl_generics #path for #type_ident #ty_generics #wc {
+                #items
+            }
+        })
+    }
+
     /// Generate an impl for a struct item
     ///
     /// The default implementation is a wrapper around [`Self::struct_items`]
@@ -94,6 +117,24 @@ pub trait ImplTrait {
                 #items
             }
         })
+    }
+
+    /// Generate enum items
+    ///
+    /// On success, this method returns the tuple `(trait_path, items)`. These
+    /// are used to generate the following implementation:
+    /// ```ignore
+    /// impl #impl_generics #trait_path for #type_ident #ty_generics #where_clause {
+    ///     #items
+    /// }
+    /// ```
+    ///
+    /// Note: this method is *only* called by the default implementation of [`Self::enum_impl`].
+    ///
+    /// Default implementation: returns an error indicating that enum expansion is not supported.
+    fn enum_items(&self, item: &ItemEnum, args: &ImplArgs) -> Result<(Toks, Toks)> {
+        let _ = (item, args);
+        Err(Error::CallSite("enum expansion not supported"))
     }
 
     /// Generate struct items
@@ -142,6 +183,20 @@ pub enum Error {
     WithSpan(Span, &'static str),
     /// Emit an error regarding path arguments
     PathArguments(&'static str),
+}
+
+impl Error {
+    /// Report via [`proc_macro_error::emit_error`].
+    pub fn emit(self, target: Span, path_args: Span) {
+        match self {
+            Error::RequireUsing => {
+                emit_error!(target, "target requires argument `using self.FIELD`")
+            }
+            Error::CallSite(msg) => emit_error!(target, msg),
+            Error::WithSpan(span, msg) => emit_error!(span, msg),
+            Error::PathArguments(msg) => emit_error!(path_args, msg),
+        }
+    }
 }
 
 /// Result type
@@ -235,9 +290,26 @@ impl ImplTraits {
     ///
     /// This attribute does not modify the item.
     /// The caller should append the result to `item` tokens.
+    ///
+    /// Errors are reported via [`proc_macro_error::emit_error`].
     pub fn expand(
         self,
         item: Toks,
+        find_impl: impl Fn(&Path) -> Option<&'static dyn ImplTrait>,
+    ) -> Toks {
+        let mut toks = Toks::new();
+        match parse2::<Item>(item) {
+            Ok(Item::Enum(item)) => toks = self.expand_enum(item, find_impl),
+            Ok(Item::Struct(item)) => toks = self.expand_struct(item, find_impl),
+            Ok(item) => emit_error!(item, "expected struct"),
+            Err(err) => emit_error!(err),
+        }
+        toks
+    }
+
+    fn expand_enum(
+        self,
+        item: ItemEnum,
         find_impl: impl Fn(&Path) -> Option<&'static dyn ImplTrait>,
     ) -> Toks {
         let ImplTraits {
@@ -245,17 +317,66 @@ impl ImplTraits {
             mut args,
         } = self;
 
-        let item = match parse2::<Item>(item) {
-            Ok(Item::Struct(item)) => item,
-            Ok(item) => {
-                emit_error!(item, "expected struct");
-                return Toks::new();
+        if !args.ignores.is_empty() {
+            let ignores = args.ignores.iter();
+            let list = quote! { #(#ignores)* };
+            emit_error!(list, "enum expansion does not currently support `ignore`",);
+            return Toks::new();
+        }
+        if let Some(mem) = args.using {
+            emit_error!(mem, "enum expansion does not currently support `using`",);
+            return Toks::new();
+        }
+
+        let mut impl_targets: Vec<(Span, _, _)> = Vec::with_capacity(targets.len());
+        for mut target in targets.drain(..) {
+            let target_span = target.span();
+            let path_args = target
+                .segments
+                .last_mut()
+                .map(|seg| std::mem::take(&mut seg.arguments))
+                .unwrap_or(PathArguments::None);
+            let target_impl = match find_impl(&target) {
+                Some(impl_) => impl_,
+                None => {
+                    emit_error!(target, "unsupported trait");
+                    return Toks::new();
+                }
+            };
+
+            if !(path_args.is_empty() || target_impl.support_path_arguments()) {
+                emit_error!(
+                    target_span,
+                    "target {} does not support path arguments",
+                    target_impl.path()
+                );
             }
-            Err(err) => {
-                emit_error!(err);
-                return Toks::new();
+
+            impl_targets.push((target.span(), target_impl, path_args));
+        }
+
+        let mut toks = Toks::new();
+
+        for (span, target, path_args) in impl_targets.drain(..) {
+            let path_args_span = path_args.span();
+            args.path_arguments = path_args;
+            match target.enum_impl(&item, &args) {
+                Ok(items) => toks.append_all(items),
+                Err(error) => error.emit(span, path_args_span),
             }
-        };
+        }
+        toks
+    }
+
+    fn expand_struct(
+        self,
+        item: ItemStruct,
+        find_impl: impl Fn(&Path) -> Option<&'static dyn ImplTrait>,
+    ) -> Toks {
+        let ImplTraits {
+            mut targets,
+            mut args,
+        } = self;
 
         let mut not_supporting_ignore = vec![];
         let mut not_supporting_using = vec![];
@@ -347,14 +468,7 @@ impl ImplTraits {
             args.path_arguments = path_args;
             match target.struct_impl(&item, &args) {
                 Ok(items) => toks.append_all(items),
-                Err(error) => match error {
-                    Error::RequireUsing => {
-                        emit_error!(span, "target requires argument `using self.FIELD`")
-                    }
-                    Error::CallSite(msg) => emit_error!(span, msg),
-                    Error::WithSpan(span, msg) => emit_error!(span, msg),
-                    Error::PathArguments(msg) => emit_error!(path_args_span, msg),
-                },
+                Err(error) => error.emit(span, path_args_span),
             }
         }
         toks